Fix Wi-Fi connection instructions
The instructions in create_connecting_image.md to connect an MBL device
to a Wi-Fi network are wrong. Just add a link to our Wi-Fi setup page.

Additionally, the extra emphasis on "DHCP server using Ethernet" doesn't
really make sense. I think it's the "with a DHCP server" bit that needs
emphasis so reword slightly and change the emphasis.