helloworld: Amend build instructions to take into account microprocessor architecture
For IOTMBL-1813 NXP 8M Mini EVK application status does not change to
RUNNING

Amend the steps to build the helloworld application to take into account
the on-board microprocessor architecture. As this add complexity to the
building process, the user is invited to use the existing shell script
that allows building the application for the correct architecture by
encapsulating the steps previously provided.